feat(api): rename package project
