+++ pom.xml
@@ -3,8 +3,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.cora</groupId>
-    <artifactId>cora-payment-api</artifactId>
+    <groupId>com.organization</groupId>
+    <artifactId>payment-api</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
@@ -39,6 +39,7 @@
         <java.faker>0.18</java.faker>
         <modelmapper.version>2.3.0</modelmapper.version>
         <springfox-swagger.version>2.9.2</springfox-swagger.version>
+        <sonar.version>3.4.0.905</sonar.version>
         <m2e.apt.activation>disabled</m2e.apt.activation>
     </properties>
 
@@ -270,6 +271,15 @@
                 </configuration>
             </plugin>
         </plugins>
+        <pluginManagement>
+	        <plugins>
+	            <plugin>
+	                <groupId>org.sonarsource.scanner.maven</groupId>
+	                <artifactId>sonar-maven-plugin</artifactId>
+	                <version>${sonar.version}</version>
+	            </plugin>
+	        </plugins>
+	    </pluginManagement>
     </build>
 
     <repositories>
