--- src/main/java/com/cora/payment/PaymentApplication.java
@@ -1,13 +1,0 @@
-package com.cora.payment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PaymentApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(PaymentApplication.class, args);
-  }
-
-}
