--- src/main/java/com/cora/payment/business/ClientBusiness.java
@@ -1,10 +1,0 @@
-package com.cora.payment.business;
-
-import com.cora.payment.entity.ClientEntity;
-
-public interface ClientBusiness {
-
-  ClientEntity save(ClientEntity client);
-
-}
-
