--- src/main/java/com/cora/payment/config/ApplicationConfig.java
@@ -1,15 +1,0 @@
-package com.cora.payment.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ApplicationConfig {
-
-  @Bean
-  public ModelMapper modelMapper() {
-    ModelMapper modelMapper = new ModelMapper();
-    return modelMapper;
-  }
-}
