--- src/main/java/com/cora/payment/config/CustomControllerAdvice.java
@@ -1,10 +1,0 @@
-package com.cora.payment.config;
-
-import org.springframework.core.annotation.Order;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice
-@Order(0)
-public class CustomControllerAdvice {
-
-}
