--- src/main/java/com/cora/payment/repository/ClientRepository.java
@@ -1,12 +1,0 @@
-package com.cora.payment.repository;
-
-import com.cora.payment.entity.ClientEntity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ClientRepository extends JpaRepository<ClientEntity, String> {
-
-}
-
