+++ src/main/java/com/organization/payment/PaymentApplication.java
@@ -1,0 +1,13 @@
+package com.organization.payment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PaymentApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(PaymentApplication.class, args);
+  }
+
+}
