+++ src/main/java/com/organization/payment/business/ClientBusiness.java
@@ -1,0 +1,10 @@
+package com.organization.payment.business;
+
+import com.organization.payment.entity.ClientEntity;
+
+public interface ClientBusiness {
+
+  ClientEntity save(ClientEntity client);
+
+}
+
