+++ src/main/java/com/organization/payment/business/impl/ClientBusinessImpl.java
@@ -1,0 +1,28 @@
+package com.organization.payment.business.impl;
+
+import com.organization.payment.business.ClientBusiness;
+import com.organization.payment.entity.ClientEntity;
+import com.organization.payment.repository.ClientRepository;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class ClientBusinessImpl implements ClientBusiness {
+
+  private final ClientRepository clientRepository;
+
+  @Autowired
+  public ClientBusinessImpl(final ClientRepository clientRepository) {
+    this.clientRepository = clientRepository;
+  }
+
+  @Override
+  public ClientEntity save(ClientEntity client) {
+    return this.clientRepository.save(client);
+  }
+
+}
+
