+++ src/main/java/com/organization/payment/config/ApplicationConfig.java
@@ -1,0 +1,15 @@
+package com.organization.payment.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ApplicationConfig {
+
+  @Bean
+  public ModelMapper modelMapper() {
+    ModelMapper modelMapper = new ModelMapper();
+    return modelMapper;
+  }
+}
