+++ src/main/java/com/organization/payment/config/CustomControllerAdvice.java
@@ -1,0 +1,10 @@
+package com.organization.payment.config;
+
+import org.springframework.core.annotation.Order;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+
+@ControllerAdvice
+@Order(0)
+public class CustomControllerAdvice {
+
+}
