+++ src/main/java/com/organization/payment/repository/ClientRepository.java
@@ -1,0 +1,12 @@
+package com.organization.payment.repository;
+
+import com.organization.payment.entity.ClientEntity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ClientRepository extends JpaRepository<ClientEntity, String> {
+
+}
+
