--- src/test/java/com/cora/payment/commons/ControllerGenericTest.java
@@ -1,8 +1,0 @@
-package com.cora.payment.commons;
-
-public abstract class ControllerGenericTest<C> extends ContextGenericTest {
-
-  protected static final String PATH_RESOURCE = "/v1/paymentapi/";
-
-}
-
