--- src/test/java/com/cora/payment/config/SpringConfigurationTest.java
@@ -1,11 +1,0 @@
-package com.cora.payment.config;
-
-import org.mockito.Mockito;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-@Configuration
-public class SpringConfigurationTest {
-
-}
