--- src/test/java/com/cora/payment/factory/AbstractFactory.java
@@ -1,9 +1,0 @@
-package com.cora.payment.factory;
-
-import java.io.Serializable;
-
-public abstract class AbstractFactory<M extends Serializable> {
-
-  abstract M simple();
-
-}
