--- src/test/java/com/cora/payment/repository/ClientRepositoryTest.java
@@ -1,31 +1,0 @@
-package com.cora.payment.repository;
-
-import com.cora.payment.commons.RepositoryGenericTest;
-import com.cora.payment.factory.ClientEntityFactory;
-import com.cora.payment.entity.ClientEntity;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-public class ClientRepositoryTest extends RepositoryGenericTest {
-
-  @Autowired
-  private ClientRepository clientRepository;
-
-  @Autowired
-  private ClientEntityFactory clientFactory;
-
-  @Test
-  public void save() {
-    ClientEntity client = this.clientFactory.simple();
-    ClientEntity clientSaved = this.clientRepository.save(client);
-
-    Assert.assertNotNull(clientSaved.getId());
-    Assert.assertEquals(client.getName(), clientSaved.getName());
-
-    this.clientRepository.delete(clientSaved);
-  }
-
-}
-
