+++ src/test/java/com/organization/payment/PaymentApplicationTest.java
@@ -1,0 +1,17 @@
+package com.organization.payment;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import com.organization.payment.PaymentApplication;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+public class PaymentApplicationTest {
+
+  @Test
+  public void main() {
+    PaymentApplication.main(new String[] {"--spring.config.location=classpath:application-test.yml"});
+  }
+
+}
