+++ src/test/java/com/organization/payment/commons/ControllerGenericTest.java
@@ -1,0 +1,8 @@
+package com.organization.payment.commons;
+
+public abstract class ControllerGenericTest<C> extends ContextGenericTest {
+
+  protected static final String PATH_RESOURCE = "/v1/paymentapi/";
+
+}
+
