+++ src/test/java/com/organization/payment/config/SpringConfigurationTest.java
@@ -1,0 +1,11 @@
+package com.organization.payment.config;
+
+import org.mockito.Mockito;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration
+public class SpringConfigurationTest {
+
+}
