+++ src/test/java/com/organization/payment/factory/AbstractFactory.java
@@ -1,0 +1,9 @@
+package com.organization.payment.factory;
+
+import java.io.Serializable;
+
+public abstract class AbstractFactory<M extends Serializable> {
+
+  abstract M simple();
+
+}
