+++ src/test/java/com/organization/payment/factory/ClientRequestDtoFactory.java
@@ -1,0 +1,17 @@
+package com.organization.payment.factory;
+
+import org.springframework.stereotype.Service;
+
+import com.organization.payment.v1.dto.ClientRequestDto;
+
+@Service
+public class ClientRequestDtoFactory extends AbstractFactory {
+
+  @Override
+  public ClientRequestDto simple() {
+    return ClientRequestDto.builder()
+      .name("Empresa 1")
+      .build();
+  }
+}
+
